requirements.txt
alembic==1.18.4
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
pycparser==3.0
PyJWT==2.11.0
PyMySQL==1.1.2
python-dotenv==1.2.1
SQLAlchemy==2.0.48
tomli==2.4.0
typing_extensions==4.15.0
Werkzeug==3.1.5
WTForms==3.2.1
